Remove dead production code; move test helpers to test files
- buildUserAuthzMaskQueryAndArgs: moved from store_authz_user_listing.go to
  store_authz_user_listing_test.go (only used by tests).
- resourceAuthzUsersBaseSQL / resourceAuthzUsersBaseArgs: moved from
  store_authz_resource_listing.go to store_authz_resource_users_test.go
  (only used by tests).
- scanUserMasks: deleted (no callers after T04 replaced the two-pass
  ResourceAuthzUsersList implementation).
- combinedMaskForUserResource: deleted (no callers; ReconcileUserResourceMasks
  and property tests use PermissionMasksForUserResource directly).
