+++ go/internal/store/sqlite/store_authz_user_listing.go
@@ -10,6 +10,10 @@ import (
 	"github.com/dtorabi/access-manager/internal/store"
 )
 
+// Note: buildUserAuthzMaskQueryAndArgs lives in store_authz_user_listing_test.go
+// — it is a test helper for verifying ground-truth masks against the materialized
+// cache and has no production callers.
+
 func userEffectivePermissionArgs(userID string) []any {
 	return []any{userID, userID}
 }
@@ -21,22 +25,6 @@ func inPlaceholders(n int) (string, error) {
 		return "", fmt.Errorf("in placeholders: n must be > 0")
 	}
 	return strings.TrimSuffix(strings.Repeat("?,", n), ","), nil
-}
-
-// buildUserAuthzMaskQueryAndArgs builds the batched mask query that returns
-// per-permission access_mask values for a user across a set of resources
-// (direct grants + group membership). Used by property tests to compare
-// against the materialized cache.
-func buildUserAuthzMaskQueryAndArgs(domainID string, resourceIDs []string, predicateArgs []any) (string, []any, error) {
-	baseSQL := `SELECT p.resource_id, p.access_mask FROM permissions p WHERE p.domain_id = ? AND p.access_mask > 0`
-	baseArgs := []any{domainID}
-	query, args, err := buildInQueryAndArgs(baseSQL, "p.resource_id", baseArgs, resourceIDs)
-	if err != nil {
-		return "", nil, err
-	}
-	query += userEffectivePermissionPredicateSQL
-	args = append(args, predicateArgs...)
-	return query, args, nil
 }
 
 func (s *Store) UserAuthzResourcesList(ctx context.Context, domainID, userID string, opts store.ListOpts) ([]store.UserAuthzResource, int64, error) {
